Refactor report popovers to use dialogs
Following up from #37638, this commit refactors `ReportPopover` and
friends to instead use `PopupDialog` and `DialogOverlay`. This results
in the report forms being always placed in the center of the screen,
instead of being anchored at whatever drawable happened to spawn it.
