+++ osu.Game/Overlays/Comments/CommentReportDialog.cs
@@ -9,14 +9,14 @@ using osu.Game.Resources.Localisation.Web;
 
 namespace osu.Game.Overlays.Comments
 {
-    public partial class ReportCommentPopover : ReportPopover<CommentReportReason>
+    public partial class CommentReportDialog : ReportDialog<CommentReportReason>
     {
         private readonly Comment comment;
 
         protected override bool IsCommentRequired(CommentReportReason reason) => reason == CommentReportReason.Other;
 
-        public ReportCommentPopover(Comment comment)
-            : base(ReportStrings.CommentTitle(comment.User?.Username ?? comment.LegacyName ?? @"Someone"), false)
+        public CommentReportDialog(Comment comment, OverlayColourProvider? colourProvider = null)
+            : base(ReportStrings.CommentTitle(comment.User?.Username ?? comment.LegacyName ?? @"Someone"), false, colourProvider)
         {
             this.comment = comment;
         }
